+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added example for running Django with auto instrumentation.
   ([#1803](https://github.com/open-telemetry/opentelemetry-python/pull/1803))
+- Added `B3SingleFormat` and `B3MultiFormat` propagators to the `opentelemetry-propagator-b3` package.
+  ([#1823](https://github.com/open-telemetry/opentelemetry-python/pull/1823))
 - Added support for OTEL_SERVICE_NAME.
   ([#1829](https://github.com/open-telemetry/opentelemetry-python/pull/1829))
 
@@ -24,6 +26,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Propagators use the root context as default for `extract` and do not modify
   the context if extracting from carrier does not work.
   ([#1811](https://github.com/open-telemetry/opentelemetry-python/pull/1811))
+- Fixed `b3` propagator entrypoint to point to `B3SingleFormat` propagator.
+  ([#1823](https://github.com/open-telemetry/opentelemetry-python/pull/1823))
+- Added `b3multi` propagator entrypoint to point to `B3MultiFormat` propagator.
+  ([#1823](https://github.com/open-telemetry/opentelemetry-python/pull/1823))
 - Improve warning when failing to decode byte attribute
   ([#1810](https://github.com/open-telemetry/opentelemetry-python/pull/1810))
 - Fixed inconsistency in parent_id formatting from the ConsoleSpanExporter